fix(nc-gui): er: refactor: consistent styling and no code rep
