+++ packages/nc-gui/components/smartsheet/expanded-form/presentors/Fields/Columns.vue
@@ -1,7 +1,5 @@
 <script lang="ts" setup>
-import { type ColumnType, isLinksOrLTAR, isVirtualCol } from 'nocodb-sdk'
-
-/* interface */
+import { type ColumnType } from 'nocodb-sdk'
 
 const props = defineProps<{
   fields: ColumnType[]
@@ -12,34 +10,11 @@ const props = defineProps<{
 
 const isLoading = toRef(props, 'isLoading')
 
-const isPublic = inject(IsPublicInj, ref(false))
+const { loadRow: _loadRow, row: _row } = useExpandedFormStoreOrThrow()
 
-/* stores */
-
-const { changedColumns, isNew, loadRow: _loadRow, row: _row } = useExpandedFormStoreOrThrow()
-
-const { isUIAllowed } = useRoles()
 const { isMobileMode } = useGlobal()
 
-/* flags */
-
-const readOnly = computed(() => !isUIAllowed('dataEdit') || isPublic.value)
-
-/* initial focus and scroll fix */
-
-const cellWrapperEl = ref()
-const mentionedCell = ref('')
-
-/* hidden fields */
-
 const showHiddenFields = ref(false)
-
-function toggleHiddenFields() {
-  showHiddenFields.value = !showHiddenFields.value
-}
-
-const shouldApplyDataCell = (column: ColumnType) =>
-  !(isBarcode(column) || isQrCode(column) || isCheckbox(column) || isRating(column) || isJSON(column))
 </script>
 
 <template>
@@ -47,83 +22,19 @@ const shouldApplyDataCell = (column: ColumnType) =>
     ref="expandedFormScrollWrapper"
     class="flex flex-col flex-grow gap-5 h-full max-h-full nc-scrollbar-thin items-center w-full p-4 xs:(px-4 pt-4 pb-2 gap-6) children:max-w-[588px] <lg:(children:max-w-[450px])"
   >
-    <div
-      v-for="(col, i) of fields"
-      v-show="!isVirtualCol(col) || !isNew || isLinksOrLTAR(col)"
-      :key="col.title"
-      :class="`nc-expand-col-${col.title}`"
-      :col-id="col.id"
-      :data-testid="`nc-expand-col-${col.title}`"
-      class="nc-expanded-form-row w-full"
-    >
-      <div
-        class="flex items-start nc-expanded-cell min-h-[37px]"
-        :class="{
-          'flex-row sm:(gap-x-2) <lg:(flex-col w-full)': !props.forceVerticalMode,
-          'flex-col w-full': props.forceVerticalMode,
-        }"
-      >
-        <div
-          class="flex items-center rounded-lg overflow-hidden"
-          :class="{
-            'w-45 <lg:(w-full px-0 mb-1) h-[37px] xs:(h-auto)': !props.forceVerticalMode,
-            'w-full px-0 mb-1 h-auto': props.forceVerticalMode,
-          }"
-        >
-          <LazySmartsheetHeaderVirtualCell
-            v-if="isVirtualCol(col)"
-            :column="col"
-            class="nc-expanded-cell-header h-full flex-none"
-          />
-          <LazySmartsheetHeaderCell v-else :column="col" class="nc-expanded-cell-header flex-none" />
-        </div>
-
-        <template v-if="isLoading">
-          <a-skeleton-input active class="h-8 flex-none <lg:!w-full lg:flex-1 !rounded-lg !overflow-hidden" size="small" />
-        </template>
-        <template v-else>
-          <SmartsheetDivDataCell
-            v-if="col.title"
-            :ref="(el: any) => { if (i) cellWrapperEl = el }"
-            class="bg-white flex-1 <lg:w-full px-1 min-h-8 flex items-center relative"
-            :class="{
-              'w-full': props.forceVerticalMode,
-              '!select-text nc-system-field bg-nc-bg-gray-extralight !text-nc-content-inverted-primary-disabled cursor-pointer':
-                isReadOnlyVirtualCell(col),
-              '!select-text nc-readonly-div-data-cell': readOnly,
-              'nc-mentioned-cell': col.id === mentionedCell,
-            }"
-            :is-data-cell="shouldApplyDataCell(col)"
-          >
-            <LazySmartsheetVirtualCell
-              v-if="isVirtualCol(col)"
-              v-model="_row.row[col.title]"
-              :column="col"
-              :read-only="readOnly"
-              :row="_row"
-            />
-
-            <LazySmartsheetCell
-              v-else
-              v-model="_row.row[col.title]"
-              :active="true"
-              :column="col"
-              :edit-enabled="true"
-              :read-only="readOnly"
-              @update:model-value="changedColumns.add(col.title)"
-            />
-          </SmartsheetDivDataCell>
-        </template>
-      </div>
-    </div>
-
+    <SmartsheetExpandedFormPresentorsFieldsColumnList
+      :store="store"
+      :fields="fields"
+      :force-vertical-mode="forceVerticalMode"
+      :is-loading="isLoading"
+    />
     <div v-if="hiddenFields.length > 0" class="flex w-full <lg:(px-1) items-center py-6">
       <div class="flex-grow h-px mr-1 bg-gray-100" />
       <NcButton
         :size="isMobileMode ? 'medium' : 'small'"
         class="flex-shrink !text-sm overflow-hidden !text-gray-500 !font-weight-500"
         type="secondary"
-        @click="toggleHiddenFields"
+        @click="showHiddenFields = !showHiddenFields"
       >
         {{ showHiddenFields ? `Hide ${hiddenFields.length} hidden` : `Show ${hiddenFields.length} hidden` }}
         {{ hiddenFields.length > 1 ? `fields` : `field` }}
@@ -131,65 +42,12 @@ const shouldApplyDataCell = (column: ColumnType) =>
       </NcButton>
       <div class="flex-grow h-px ml-1 bg-gray-100" />
     </div>
-
-    <template v-if="hiddenFields.length > 0 && showHiddenFields">
-      <div
-        v-for="(col, i) of hiddenFields"
-        v-show="isFormula(col) || !isVirtualCol(col) || !isNew || isLinksOrLTAR(col)"
-        :key="`${col.id}-${col.title}`"
-        :class="`nc-expand-col-${col.title}`"
-        :data-testid="`nc-expand-col-${col.title}`"
-        class="nc-expanded-form-row w-full"
-      >
-        <div class="flex items-start flex-row sm:(gap-x-2) <lg:(flex-col w-full) nc-expanded-cell min-h-[37px]">
-          <div class="w-45 <lg:(w-full px-0) h-[37px] xs:(h-auto) flex items-center rounded-lg overflow-hidden">
-            <LazySmartsheetHeaderVirtualCell
-              v-if="isVirtualCol(col)"
-              :column="col"
-              is-hidden-col
-              class="nc-expanded-cell-header flex-none"
-            />
-
-            <LazySmartsheetHeaderCell v-else :column="col" is-hidden-col class="nc-expanded-cell-header flex-none" />
-          </div>
-
-          <template v-if="isLoading">
-            <a-skeleton-input active class="h-8 flex-none <lg:!w-full lg:flex-1 !rounded-lg !overflow-hidden" size="small" />
-          </template>
-          <template v-else>
-            <LazySmartsheetDivDataCell
-              v-if="col.title"
-              :ref="(el: any) => { if (i) cellWrapperEl = el }"
-              class="bg-white flex-1 <lg:w-full px-1 min-h-8 flex items-center relative"
-              :class="{
-                'w-full': props.forceVerticalMode,
-                '!select-text nc-system-field bg-nc-bg-gray-extralight !text-nc-content-inverted-primary-disabled cursor-pointer':
-                  isReadOnlyVirtualCell(col),
-                '!select-text nc-readonly-div-data-cell': readOnly,
-                'nc-mentioned-cell': col.id === mentionedCell,
-              }"
-            >
-              <LazySmartsheetVirtualCell
-                v-if="isVirtualCol(col)"
-                v-model="_row.row[col.title]"
-                :column="col"
-                :read-only="readOnly"
-                :row="_row"
-              />
-
-              <LazySmartsheetCell
-                v-else
-                v-model="_row.row[col.title]"
-                :active="true"
-                :column="col"
-                :edit-enabled="true"
-                :read-only="readOnly"
-                @update:model-value="changedColumns.add(col.title)"
-              />
-            </LazySmartsheetDivDataCell>
-          </template>
-        </div>
-      </div>
-    </template>
+    <SmartsheetExpandedFormPresentorsFieldsColumnList
+      v-if="hiddenFields.length > 0 && showHiddenFields"
+      :store="store"
+      :fields="hiddenFields"
+      :force-vertical-mode="forceVerticalMode"
+      :is-loading="isLoading"
+    />
   </div>
 </template>
